pyproject.toml
[build-system]
requires = ["setuptools>=42"]
build-backend = "setuptools.build_meta"

[project]
name = "SamsScaleHouse"
version = "0.1.0"
description = "Scrap POS/POP with scale integration"
requires-python = ">=3.10"
dependencies = []

[project.optional-dependencies]
serial = ["pyserial>=3.5"]
dev = ["pytest>=7.0"]

[tool.pytest.ini_options]
testpaths = ["tests"]
